remove test extensions
