+++ articles/wagtail_hooks.py
@@ -1,79 +1,7 @@
-from wagtail.admin.rich_text.converters.html_to_contentstate import (
-    BlockElementHandler,
-    InlineStyleElementHandler,
-)
 import wagtail.admin.rich_text.editors.draftail.features as draftail_features
 from wagtail.core import hooks
 
-from .rich_text import stock_entity_decorator, StockEntityElementHandler, AnchorEntityElementHandler, anchor_entity_decorator
-
-
-@hooks.register('register_rich_text_features')
-def register_strikethrough_feature(features):
-    feature_name = 'strikethrough'
-    type_ = 'STRIKETHROUGH'
-    tag = 's'
-
-    control = {
-        'type': type_,
-        'label': 'S',
-        'description': 'Strikethrough',
-    }
-
-    features.register_editor_plugin(
-        'draftail', feature_name, draftail_features.InlineStyleFeature(control),
-    )
-
-    db_conversion = {
-        'from_database_format': {tag: InlineStyleElementHandler(type_)},
-        'to_database_format': {'style_map': {type_: tag}},
-    }
-
-    features.register_converter_rule('contentstate', feature_name, db_conversion)
-
-
-@hooks.register('register_rich_text_features')
-def register_blockquote_feature(features):
-    feature_name = 'blockquote'
-    type_ = 'BLOCKQUOTE'
-    tag = 'blockquote'
-
-    control = {
-        'type': type_,
-        'label': '❝',
-        'description': 'Blockquote',
-        'element': 'blockquote',
-    }
-
-    features.register_editor_plugin(
-        'draftail', feature_name, draftail_features.BlockFeature(control)
-    )
-
-    features.register_converter_rule('contentstate', feature_name, {
-        'from_database_format': {tag: BlockElementHandler(type_)},
-        'to_database_format': {'block_map': {type_: tag}},
-    })
-
-
-@hooks.register('register_rich_text_features')
-def register_stock_feature(features):
-    feature_name = 'stock'
-    type_ = 'STOCK'
-
-    control = {
-        'type': type_,
-        'label': '$',
-        'description': 'Stock',
-    }
-
-    features.register_editor_plugin(
-        'draftail', feature_name, draftail_features.EntityFeature(control)
-    )
-
-    features.register_converter_rule('contentstate', feature_name, {
-        'from_database_format': {'span[data-stock]': StockEntityElementHandler(type_)},
-        'to_database_format': {'entity_decorators': {type_: stock_entity_decorator}},
-    })
+from .rich_text import AnchorEntityElementHandler, anchor_entity_decorator
 
 
 @hooks.register('register_rich_text_features')
@@ -95,7 +23,7 @@ def register_rich_text_features(features):
     features.register_editor_plugin(
         'draftail', feature_name, draftail_features.EntityFeature(
             control,
-            js=['stock.js'],
+            js=['admin/anchor.js'],
         )
     )
